Fix code type

+++ compiler/console/console.hpp
@@ -69,7 +69,7 @@ namespace Console {
         extern size_t endLine;
         extern size_t endColumn;
 
-        extern size_t code;
+        extern int code;
     }
 
     // Keep track of general report statistics
